refractor(changed folders with UI components.)
