+++ src/components/landing/AboutApp.tsx
