+++ src/components/landing/Hero.tsx
@@ -1,39 +1,36 @@
-import { GitHubLogoIcon } from "@radix-ui/react-icons";
+import Image from "next/image";
+import React from "react";
+import macbook_mockup from "../../assets/macbook_mockup.png";
+import cube from "../../assets/cube.png";
+import sphere from "../../assets/sphere.png";
+import { Button } from "@/components/ui/button"
 
-import { Button, buttonVariants } from "@/components/ui/button";
-
-import { HeroCards } from "./HeroCards";
-
-export const Hero = () => {
+const Hero = () => {
   return (
-    <section className="container grid lg:grid-cols-2 place-items-center py-20 md:py-32 gap-10">
-      <div className="text-center lg:text-start space-y-6">
-        <main className="text-5xl md:text-6xl font-bold">
-          <h1 className="inline">
-            <span className="inline bg-gradient-to-r from-[#61DAFB]  to-[#80B3FF] text-transparent bg-clip-text">
-              Eventownik
-            </span>
-            , zróbmy razem wydarzenie
-          </h1>{" "}
-        </main>
-
-        <p className="text-xl text-muted-foreground md:w-10/12 mx-auto lg:mx-0">
-          Prosty i niezawodny sposób na zorganizowanie wydarzenia. Wyślij link i
-          czekaj!
-        </p>
-
-        <div className="space-y-4 md:space-y-0 md:space-x-4">
-          <Button className="w-full md:w-1/3">Dołącz do nas</Button>
+    <section>
+      <div className="p-12  pb-20 xl:pb-40 flex flex-col xl:flex-row xl:justify-between xl:items-center mt-8 xl:mt-24 ">
+        <div className="mb-12 xl:mb-0">
+          <h1 className="text-3xl md:text-5xl 2xl:text-7xl text-primary-foreground mb-8 h-3/4 text order-2 xl:order-1 ">
+            Eventownik - zróbmy razem
+            <br />
+            wydarzenie!
+          </h1>
+          <Button variant="secondary" className={'bg-secondary py-8 px-8 rounded-md font-bold text-lg'}>Utwórz wydarzenie</Button>
+        </div>
+        <div className="w-full xl:w-auto h-100 xl:h-auto mr-24 2xl:mr-56 order-1 xl:order-2 flex xl:flex-none items-center justify-center relative">
+          <Image alt="Laptop" src={macbook_mockup} className="block" />
+          <div className="xl:absolute top-0 left-0 right-0 bottom-0 flex items-center justify-center">
+            <div className="xl:absolute top-0 -left-48">
+              <Image alt="Sphere" src={sphere} className="hidden 2xl:block" />
+            </div>
+            <div className="xl:absolute bottom-0 -right-48">
+              <Image alt="Cube" src={cube} className="hidden 2xl:block" />
+            </div>
+          </div>
         </div>
       </div>
-
-      {/* Hero cards sections */}
-      <div className="z-10">
-        <HeroCards />
-      </div>
-
-      {/* Shadow effect */}
-      <div className="shadow"></div>
     </section>
   );
 };
+
+export default Hero;
