--- src/components/landing/HeroCards.tsx
@@ -1,127 +1,0 @@
-import { GitHubLogoIcon } from "@radix-ui/react-icons";
-import { Check, Linkedin } from "lucide-react";
-import Image from "next/image";
-
-import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
-import { Badge } from "@/components/ui/badge";
-import { Button, buttonVariants } from "@/components/ui/button";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardFooter,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card";
-
-import { LightBulbIcon } from "./Icons";
-
-export const HeroCards = () => {
-  return (
-    <div className="hidden lg:flex flex-row flex-wrap gap-8 relative w-[700px] h-[500px]">
-      {/* Testimonial */}
-      <Card className="absolute w-[340px] -top-[15px] drop-shadow-xl shadow-black/10 dark:shadow-white/10">
-        <CardHeader className="flex flex-row items-center gap-4 pb-2">
-          <Avatar>
-            <AvatarImage
-              alt=""
-              src="https://cms.solvro.pl/assets/77255d90-0ec0-4219-9c81-707cc50babc4?key=member"
-            />
-            <AvatarFallback>SH</AvatarFallback>
-          </Avatar>
-
-          <div className="flex flex-col">
-            <CardTitle className="text-lg">Dawid Linek</CardTitle>
-            <CardDescription>@dawid_linek</CardDescription>
-          </div>
-        </CardHeader>
-
-        <CardContent>Żałuję, że wcześniej o tym nie wiedziałem!</CardContent>
-      </Card>
-
-      {/* Team */}
-      <Card className="absolute right-[20px] top-4 w-80 flex flex-col justify-center items-center drop-shadow-xl shadow-black/10 dark:shadow-white/10">
-        <CardHeader className="mt-8 flex justify-center items-center pb-2">
-          <Image
-            src="https://cms.solvro.pl/assets/52c20245-1811-4dd3-b389-f9be5948e642?key=member"
-            alt="user avatar"
-            width={200}
-            height={200}
-            className="absolute grayscale-[0%] -top-12 rounded-full w-24 h-24 aspect-square object-cover"
-          />
-          <CardTitle className="text-center">Jakub Karbowski</CardTitle>
-          <CardDescription className="font-normal text-primary">
-            Frontend Developer
-          </CardDescription>
-        </CardHeader>
-
-        <CardContent className="text-center pb-2">
-          <p>Praca nad Eventownikiem to sama przyjemność.</p>
-        </CardContent>
-
-        <CardFooter>
-          <div>
-            <a
-              rel="noreferrer noopener"
-              href="https://github.com/karbowskijakub"
-              target="_blank"
-              className={buttonVariants({
-                variant: "ghost",
-                size: "sm",
-              })}
-            >
-              <span className="sr-only">Github icon</span>
-              <GitHubLogoIcon className="w-5 h-5" />
-            </a>
-            <a
-              rel="noreferrer noopener"
-              href="https://www.linkedin.com/in/jakub-karbowski-48a254217"
-              target="_blank"
-              className={buttonVariants({
-                variant: "ghost",
-                size: "sm",
-              })}
-            >
-              <span className="sr-only">Linkedin icon</span>
-              <Linkedin size="20" />
-            </a>
-          </div>
-        </CardFooter>
-      </Card>
-
-      {/* Pricing */}
-      <Card className="absolute top-[150px] left-[50px] w-72  drop-shadow-xl shadow-black/10 dark:shadow-white/10">
-        <CardHeader>
-          <CardTitle className="flex item-center justify-between">
-            Darmowo
-          </CardTitle>
-          <div>
-            <span className="text-3xl font-bold">0zł</span>
-            <span className="text-muted-foreground"> /miesiąc</span>
-          </div>
-
-          <CardDescription>
-            Eventownik był, jest i będzie darmowy.
-          </CardDescription>
-        </CardHeader>
-
-        <hr className="w-4/5 m-auto mb-4" />
-
-        <CardFooter className="flex">
-          <div className="space-y-4">
-            {[
-              "Nielimitowane wydarzenia",
-              "Nielimitowany czas",
-              "Nielimitowani uczestnicy",
-            ].map((benefit: string) => (
-              <span key={benefit} className="flex">
-                <Check className="text-green-500" />{" "}
-                <h3 className="ml-2">{benefit}</h3>
-              </span>
-            ))}
-          </div>
-        </CardFooter>
-      </Card>
-    </div>
-  );
-};
